--- src/components/PeoplePage.tsx
@@ -1,116 +1,0 @@
-import { Loader } from './Loader';
-import { useEffect, useState } from 'react';
-import { getPeople } from '../api';
-import { Person } from '../types/Person';
-import { useParams } from 'react-router-dom';
-
-import { PersonLink } from './PersonLink';
-import classNames from 'classnames';
-
-export const PeoplePage = () => {
-  const [isLoading, setIsLoading] = useState<boolean>(true);
-  const [people, setPeople] = useState<Person[]>([]);
-  const [hasError, setHasError] = useState<boolean>(false);
-
-  const loadPersons = async () => {
-    try {
-      setIsLoading(true);
-      setHasError(false);
-      const result = await getPeople();
-
-      setPeople(result);
-    } catch {
-      setHasError(true);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    loadPersons();
-  }, []);
-
-  const { slug } = useParams();
-
-  return (
-    <>
-      <h1 className="title">People Page</h1>
-
-      <div className="block">
-        <div className="box table-container">
-          {isLoading && <Loader />}
-
-          {!isLoading && hasError && (
-            <p data-cy="peopleLoadingError" className="has-text-danger">
-              Something went wrong
-            </p>
-          )}
-
-          {!isLoading && !hasError && people.length === 0 && (
-            <p data-cy="noPeopleMessage">There are no people on the server</p>
-          )}
-
-          {!isLoading && !hasError && people.length > 0 && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => {
-                  const mother = people.find(p => p.name === person.motherName);
-                  const father = people.find(p => p.name === person.fatherName);
-
-                  return (
-                    <tr
-                      key={person.slug}
-                      data-cy="person"
-                      className={classNames({
-                        'has-background-warning': person.slug === slug,
-                      })}
-                    >
-                      <td>
-                        <PersonLink person={person} />
-                      </td>
-                      <td>{person.sex}</td>
-                      <td>{person.born}</td>
-                      <td>{person.died}</td>
-                      <td>
-                        {person.motherName === null ? (
-                          '-'
-                        ) : mother ? (
-                          <PersonLink person={mother} />
-                        ) : (
-                          person.motherName
-                        )}
-                      </td>
-                      <td>
-                        {person.fatherName === null ? (
-                          '-'
-                        ) : father ? (
-                          <PersonLink person={father} />
-                        ) : (
-                          person.fatherName
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          )}
-        </div>
-      </div>
-    </>
-  );
-};
