+++ src/pages/PeoplePage.tsx
@@ -1,0 +1,58 @@
+import { Loader } from '../components/Loader';
+import { useEffect, useState } from 'react';
+import { getPeople } from '../api';
+import { Person } from '../types';
+import { useParams } from 'react-router-dom';
+import { PeopleTable } from '../components/PeopleTable';
+
+export const PeoplePage = () => {
+  const [isLoading, setIsLoading] = useState<boolean>(true);
+  const [people, setPeople] = useState<Person[]>([]);
+  const [hasError, setHasError] = useState<boolean>(false);
+
+  const loadPersons = async () => {
+    try {
+      setIsLoading(true);
+      setHasError(false);
+      const result = await getPeople();
+
+      setPeople(result);
+    } catch {
+      setHasError(true);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    loadPersons();
+  }, []);
+
+  const { slug } = useParams();
+
+  return (
+    <>
+      <h1 className="title">People Page</h1>
+
+      <div className="block">
+        <div className="box table-container">
+          {isLoading && <Loader />}
+
+          {!isLoading && hasError && (
+            <p data-cy="peopleLoadingError" className="has-text-danger">
+              Something went wrong
+            </p>
+          )}
+
+          {!isLoading && !hasError && people.length === 0 && (
+            <p data-cy="noPeopleMessage">There are no people on the server</p>
+          )}
+
+          {!isLoading && !hasError && people.length > 0 && (
+            <PeopleTable people={people} selectedSlug={slug} />
+          )}
+        </div>
+      </div>
+    </>
+  );
+};
